requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
meilisearch==0.34.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
requests==2.31.0
transformers==4.33.2
torch==2.0.1
numpy==1.24.3 